requirements.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.3.2
pylint==1.7.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
PyPDF2==1.26.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
